Move `.group` and `.stack` to `compat` layer.
