+++ src/charts/text-filter-widget.ts
@@ -54,13 +54,6 @@ export class TextFilterWidget extends BaseMixin {
             },
         });
 
-        // @ts-ignore, signature is different in BaseMixin
-        this.group(() => {
-            throw new Error(
-                'the group function on textFilterWidget should never be called, please report the issue'
-            );
-        });
-
         this.anchor(parent, chartGroup);
     }
